requirements.txt
Django<1.9.0
MySQL-python<2.0
djangorestframework<3.2

django_extensions<1.6
ipython<3.2